Profundas refactorizaciones y correcciones a usuarios, pistas y reservas
- Se ha sustituido el uso de la variable request del controlador por una
llamada al método getRequest(), pues la primera será obsoleta en el futuro
cercano.
- Se han mejorado las comprobaciones de acceso en el controlador de usuarios,
de forma que un deportista ya no podrá ver ni editar otros usuarios aunque
sepa su ID y construya el URL correspondiente.
- Se han armonizado las vistas de reservas, pistas y usuarios en estilo de
código y visual.
- Se ha movido la acción index de usuario al controlador de páginas, pues
simplemente mostraba una página estática sin lógica. Esto simplifica el controlador
de usuarios y le da un propósito al controlador de páginas. Ahora el método
index de usuario muestra todos los usuarios del sistema, como en el resto de CRUD.
Además, esta página actúa de índice para usuarios no registrados, siendo ésta
lo único que pueden ver sin iniciar sesión.
- Se han corregido los enlaces de la barra de navegación.

+++ src/Controller/UsuarioController.php
@@ -2,7 +2,7 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
-use Cake\Event\Event;
+use Cake\Datasource\Exception\RecordNotFoundException;
 
 /**
  * Usuario Controller
@@ -14,12 +14,14 @@ use Cake\Event\Event;
 class UsuarioController extends AppController
 {
     /**
-     * Index method
-     *
-     * @return void
+     * {@inheritDoc}
      */
+    public function initialize()
+    {
+        parent::initialize();
 
-
+        $this->Auth->allow(['register', 'login']);
+    }
 
     /**
      * Comprueba que el usuario conectado tenga los privilegios suficientes para interactuar
@@ -31,53 +33,41 @@ class UsuarioController extends AppController
      * @return bool Verdadero si se le debe de conceder acceso a la acción al usuario, falso en
      * caso contrario.
      */
-    public function beforeFilter(Event $event)
-    {
-        parent::beforeFilter($event);
-        $this->Auth->allow('index');
-        $this->Auth->allow('register');
-        $this->Auth->allow('login');
-        $this->Auth->allow('logout');
-    }
-
     public function isAuthorized($user)
     {
-        // Los usuarios no administradores solo tienen acceso a las acciones index y logout.
-        // De otro modo, el proceso de conexión desembocaría en un bucle infinito de redirecciones,
-        // y los usuarios no se podrían desconectar
-        return in_array($this->request->getParam('action'), ['register','index', 'logout', 'edit']) ||
+        // Los usuarios no administradores solo pueden ver su perfil, editarlo
+        // y desconectarse
+        return in_array($this->getRequest()->getParam('action'), ['view', 'edit', 'logout']) ||
                $user['rol'] === 'administrador';
-
     }
 
-    
-    public function index(){
-
-    }
-
-    public function listar()
+    /**
+     * Index method
+     *
+     * @return void
+     */
+    public function index()
     {
-
         $usuario = $this->paginate($this->Usuario);
-        $this->set('usuario', $usuario);
-
+        $this->set('usuarios', $usuario);
     }
-
 
     /**
      * Inicia sesión
      *
-     * @return void
+     * @return \Cake\Http\Response|null Redirects on successful login, renders view otherwise.
      */
     public function login()
     {
         if ($this->Auth->user() !== null) {
             $this->Flash->success(__('Ya estás conectado como {0}.', $this->Auth->user('username')));
-            $this->redirect($this->referer(['controller' => $this->getName(), 'action' => 'index'], true));
-        } elseif ($this->request->is('post')) {
+
+            return $this->redirect($this->referer(['controller' => 'Pages', 'action' => 'display', 'index'], true));
+        } elseif ($this->getRequest()->is('post')) {
             $usuario = $this->Auth->identify();
             if ($usuario) {
                 $this->Auth->setUser($usuario);
+
                 return $this->redirect($this->Auth->redirectUrl());
             } else {
                 $this->Flash->error(__('Alguna credencial es incorrecta. Por favor, revisa que no hayas escrito algo mal e inténtalo de nuevo.'));
@@ -85,6 +75,11 @@ class UsuarioController extends AppController
         }
     }
 
+    /**
+     * Cierra sesión
+     *
+     * @return \Cake\Http\Response Redirects to index.
+     */
     public function logout()
     {
         return $this->redirect($this->Auth->logout());
@@ -93,84 +88,157 @@ class UsuarioController extends AppController
     /**
      * Registra a un usuario
      *
-     * @return void
+     * @return \Cake\Http\Response|null
      */
     public function register()
     {
         $usuario = $this->Usuario->newEntity();
-        $this->viewBuilder();
-        
-        if ($this->request->is('post')) {
-            $data = $this->request->getData();
+
+        if ($this->getRequest()->is('post')) {
+            /** @var array */
+            $data = $this->getRequest()->getData();
+
             $data['rol'] = 'deportista';
             $data['esSocio'] = '0';
-            $data['password'] = $this->hashPassword($data['password']);
+            if (isset($data['password']) && is_string($data['password'])) {
+                if (empty($data['password'])) {
+                    unset($data['password']);
+                } else {
+                    $data['password'] = $this->hashPassword($data['password']);
+                }
+            }
+
             $usuario = $this->Usuario->patchEntity($usuario, $data);
             if ($this->Usuario->save($usuario)) {
-                $this->Flash->success(__('El usuario ha sigo registrado'));
+                $this->Flash->success(__("¡Bienvenido a PadeGest, {0}!", $usuario->nombre_completo));
+                $this->Auth->setUser($usuario);
 
-                return $this->redirect(['action' => 'login']);
+                return $this->redirect($this->Auth->redirectUrl());
+            } else {
+                $this->Flash->error(__('Ha ocurrido un error al crear el nuevo usuario. Por favor, inténtalo de nuevo.'));
             }
-            $this->Flash->error(__('El usuario no se ha podido registrar, inténtalo de nuevo'));
+        } else {
+            $usuario = $this->Usuario->newEntity();
         }
+
+        $this->set(compact('usuario'));
     }
 
+    /**
+     * Add method
+     *
+     * @return \Cake\Http\Response|null Redirects on successful add, renders view otherwise.
+     */
     public function add()
     {
         $usuario = $this->Usuario->newEntity();
-        $this->viewBuilder();
-        
-        if ($this->request->is('post')) {
-            $usuario = $this->Usuario->patchEntity($usuario, $this->request->getData());
-            if ($this->Usuario->save($usuario)) {
-                $this->Flash->success(__('El usuario ha sigo registrado'));
-                return $this->redirect(['action' => 'listar']);
+
+        if ($this->getRequest()->is('post')) {
+            $data = $this->getRequest()->getData();
+
+            if (isset($data['password']) && is_string($data['password'])) {
+                if (empty($data['password'])) {
+                    unset($data['password']);
+                } else {
+                    $data['password'] = $this->hashPassword($data['password']);
+                }
             }
-            $this->Flash->error(__('El usuario no se ha podido registrar, inténtalo de nuevo'));
-        } 
+
+            $usuario = $this->Usuario->patchEntity($usuario, $data);
+            if ($this->Usuario->save($usuario)) {
+                $this->Flash->success(__('El usuario ha sido creado con éxito.'));
+
+                return $this->redirect(['action' => 'index']);
+            }
+            $this->Flash->error(__('Ha ocurrido un error al crear el usuario. Por favor, inténtalo de nuevo.'));
+        }
+
+        $this->set(compact('usuario'));
     }
 
     /**
      * View method
      *
      * @param string|null $id Usuario id.
-     * @return void
-     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
+     * @return \Cake\Http\Response|null Redirects on error, renders view otherwise.
      */
     public function view($id = null)
     {
-        $usuario = $this->Usuario->get($id);
+        $getOptions = [];
 
-        $this->set('usuario', $usuario);
+        // Los usuarios no administradores solo ven su perfil
+        if ($this->Auth->user('rol') !== 'administrador') {
+            $getOptions += [
+                'conditions' => ['id =' => $this->Auth->user('id')]
+            ];
+        }
+
+        try {
+            $usuario = $this->Usuario->get($id, $getOptions);
+        } catch (RecordNotFoundException $_) {
+            $this->Flash->error(__('El {0} especificado no existe o no tienes permiso para verlo.', __('usuario')));
+
+            return $this->redirect($this->referer(['controller' => 'Pages', 'action' => 'display', 'index'], true));
+        }
+
+        $this->set(compact('usuario'));
     }
 
     /**
      * Edit method
      *
      * @param string|null $id Usuario id.
-     * @return \Cake\Http\Response|null Redirects on successful edit, renders view otherwise.
-     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
+     * @return \Cake\Http\Response|null Redirects on edit, renders view otherwise.
      */
     public function edit($id = null)
     {
-        $usuario = $this->Usuario->get($id);
-        $var = $usuario->password;
-        $usuario->__set('password', '');
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $data = $this->request->getData();
-            if($data['password'] == ''){
-                $data['password'] = $var;
-            }else{
+        $getOptions = [];
+
+        // Los usuarios no administradores solo pueden editar su perfil
+        if ($this->Auth->user('rol') !== 'administrador') {
+            $getOptions += [
+                'conditions' => ['id =' => $this->Auth->user('id')]
+            ];
+        }
+
+        try {
+            $usuario = $this->Usuario->get($id, $getOptions);
+        } catch (RecordNotFoundException $_) {
+            $this->Flash->error(__('El {0} especificado no existe o no tienes permiso para verlo.', __('usuario')));
+
+            return $this->redirect($this->referer(['controller' => 'Pages', 'action' => 'display', 'index'], true));
+        }
+
+        if ($this->getRequest()->is(['patch', 'post', 'put'])) {
+            $data = $this->getRequest()->getData();
+
+            if (empty($data['password'])) {
+                $data['password'] = $usuario->password;
+            } else {
                 $data['password'] = $this->hashPassword($data['password']);
             }
+
             $usuario = $this->Usuario->patchEntity($usuario, $data);
             if ($this->Usuario->save($usuario)) {
-                $this->Flash->success(__('The usuario has been saved.'));
+                $this->Flash->success(__($this->Auth->user('id') == $usuario->id ? 'Has editado tu perfil.' : 'El usuario ha sido editado.'));
 
-                return $this->redirect(['action' => 'listar']);
+                // Actualizar los datos de usuario guardados en la sesión,
+                // para que los cambios en p. ej. el nombre de usuario se visualicen
+                // inmediatamente
+                if ($this->Auth->user('id') == $usuario->id) {
+                    $this->Auth->setUser($usuario);
+                }
+
+                if ($this->Auth->user('rol') !== 'administrador') {
+                    return $this->redirect(['action' => 'view', $usuario->id]);
+                } else {
+                    return $this->redirect(['action' => 'index']);
+                }
             }
-            $this->Flash->error(__('The usuario could not be saved. Please, try again.'));
+
+            $this->Flash->error(__('Ha ocurrido un error al realizar la operación solicitada. Por favor, vuélvelo a intentar más tarde.'));
         }
+
         $this->set(compact('usuario'));
     }
 
@@ -178,24 +246,39 @@ class UsuarioController extends AppController
      * Delete method
      *
      * @param string|null $id Usuario id.
-     * @return \Cake\Http\Response|null Redirects to index.
-     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
+     * @return \Cake\Http\Response Redirecciona al índice.
      */
     public function delete($id = null)
     {
-        
-        $this->request->allowMethod(['post', 'delete']);
-        $usuario = $this->Usuario->get($id);
-        if ($this->Usuario->delete($usuario)) {
-            $this->Flash->success(__('The usuario has been deleted.'));
-        } else {
-            $this->Flash->error(__('The usuario could not be deleted. Please, try again.'));
+        $this->getRequest()->allowMethod(['post', 'delete']);
+
+        try {
+            $usuario = $this->Usuario->get($id);
+
+            if ($usuario->id == $this->Auth->user('id')) {
+                $this->Flash->error(__('No puedes borrar tu propio usuario.'));
+            } else {
+                if ($this->Usuario->delete($usuario)) {
+                    $this->Flash->success(__('{0} borrado con éxito.', __('Usuario')));
+                } else {
+                    $this->Flash->error(__('Ha ocurrido un error al realizar la operación solicitada. Por favor, vuélvelo a intentar más tarde.'));
+                }
+            }
+        } catch (RecordNotFoundException $_) {
+            // Ignorar el error
+            $this->Flash->success(__('{0} borrado con éxito.', __('Usuario')));
         }
 
-        return $this->redirect(['action' => 'listar']);
-  
+        return $this->redirect(['action' => 'index']);
     }
 
+    /**
+     * Calcula el resumen de una contraseña, listo para asignar al correspondiente atributo
+     * del modelo de datos del usuario.
+     *
+     * @param string $password La contraseña de la que calcular su resumen.
+     * @return string El resumen de la contraseña.
+     */
     private function hashPassword($password)
     {
         return $this->Auth->getAuthenticate('Form')->passwordHasher()->hash($password);
